+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Skip update of ldcache in containers without ldconfig. The .so.SONAME symlinks are still created.
 * Normalize ldconfig path on use. This automatically adjust the ldconfig setting applied to ldconfig.real on systems where this exists.
+* Include `nvidia/nvoptix.bin` in list of graphics mounts.
 
 * [libnvidia-container] Fix device permission check when using cgroupv2 (fixes #227)