+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 * Normalize ldconfig path on use. This automatically adjust the ldconfig setting applied to ldconfig.real on systems where this exists.
 * Include `nvidia/nvoptix.bin` in list of graphics mounts.
 * Include `vulkan/icd.d/nvidia_layers.json` in list of graphics mounts.
-
+* Add support for `--library-search-paths` to `nvidia-ctk cdi generate` command.
 
 * [libnvidia-container] Fix device permission check when using cgroupv2 (fixes #227)